Add per query cycle handlers and use them for representability
