+++ compiler/rustc_middle/src/ty/inhabitedness/mod.rs
@@ -61,9 +61,8 @@ pub(crate) fn provide(providers: &mut Providers) {
 /// requires calling [`InhabitedPredicate::instantiate`]
 fn inhabited_predicate_adt(tcx: TyCtxt<'_>, def_id: DefId) -> InhabitedPredicate<'_> {
     if let Some(def_id) = def_id.as_local() {
-        if matches!(tcx.representability(def_id), ty::Representability::Infinite(_)) {
-            return InhabitedPredicate::True;
-        }
+        // Check for a cycle
+        let _ = tcx.representability(def_id);
     }
     let adt = tcx.adt_def(def_id);
     InhabitedPredicate::any(
