+++ compiler/rustc_middle/src/util/mod.rs
@@ -4,5 +4,6 @@ pub mod bug;
 pub struct Providers {
     pub queries: crate::queries::Providers,
     pub extern_queries: crate::queries::ExternProviders,
+    pub query_cycle_handlers: crate::queries::CycleHandlers,
     pub hooks: crate::hooks::Providers,
 }
