+++ tests/ui/enum-discriminant/issue-72554.rs
@@ -3,13 +3,11 @@ use std::collections::BTreeSet;
 #[derive(Hash)]
 pub enum ElemDerived {
     //~^ ERROR recursive type `ElemDerived` has infinite size
-    //~| ERROR cycle detected
-    A(ElemDerived)
+    A(ElemDerived),
 }
 
-
 pub enum Elem {
-    Derived(ElemDerived)
+    Derived(ElemDerived),
 }
 
 pub struct Set(BTreeSet<Elem>);
