+++ tests/ui/infinite/infinite-struct.rs
@@ -1,7 +1,5 @@
 struct Take(Take);
 //~^ ERROR has infinite size
-//~| ERROR cycle
-//~| ERROR reached the recursion limit finding the struct tail for `Take`
 
 // check that we don't hang trying to find the tail of a recursive struct (#79437)
 fn foo() -> Take {
@@ -9,7 +7,8 @@ fn foo() -> Take {
 }
 
 // mutually infinite structs
-struct Foo { //~ ERROR has infinite size
+struct Foo {
+    //~^ ERROR has infinite size
     x: Bar<Foo>,
 }
