+++ tests/ui/infinite/infinite-tag-type-recursion.rs
@@ -1,5 +1,9 @@
-enum MList { Cons(isize, MList), Nil }
-//~^ ERROR recursive type `MList` has infinite size
-//~| ERROR cycle
+enum MList {
+    //~^ ERROR recursive type `MList` has infinite size
+    Cons(isize, MList),
+    Nil,
+}
 
-fn main() { let a = MList::Cons(10, MList::Cons(11, MList::Nil)); }
+fn main() {
+    let a = MList::Cons(10, MList::Cons(11, MList::Nil));
+}
