+++ tests/ui/infinite/infinite-tag-type-recursion.stderr
@@ -1,25 +1,17 @@
 error[E0072]: recursive type `MList` has infinite size
   --> $DIR/infinite-tag-type-recursion.rs:1:1
    |
-LL | enum MList { Cons(isize, MList), Nil }
-   | ^^^^^^^^^^               ----- recursive without indirection
+LL | enum MList {
+   | ^^^^^^^^^^
+LL |
+LL |     Cons(isize, MList),
+   |                 ----- recursive without indirection
    |
 help: insert some indirection (e.g., a `Box`, `Rc`, or `&`) to break the cycle
    |
-LL | enum MList { Cons(isize, Box<MList>), Nil }
-   |                          ++++     +
+LL |     Cons(isize, Box<MList>),
+   |                 ++++     +
 
-error[E0391]: cycle detected when computing when `MList` needs drop
-  --> $DIR/infinite-tag-type-recursion.rs:1:1
-   |
-LL | enum MList { Cons(isize, MList), Nil }
-   | ^^^^^^^^^^
-   |
-   = note: ...which immediately requires computing when `MList` needs drop again
-   = note: cycle used when computing whether `MList` needs drop
-   = note: see https://rustc-dev-guide.rust-lang.org/overview.html#queries and https://rustc-dev-guide.rust-lang.org/query.html for more information
+error: aborting due to 1 previous error
 
-error: aborting due to 2 previous errors
-
-Some errors have detailed explanations: E0072, E0391.
-For more information about an error, try `rustc --explain E0072`.
+For more information about this error, try `rustc --explain E0072`.
