+++ tests/ui/query-system/query-cycle-printing-issue-151226.rs
@@ -1,8 +1,6 @@
 struct A<T>(std::sync::OnceLock<Self>);
 //~^ ERROR recursive type `A` has infinite size
-//~| ERROR cycle detected when computing layout of `A<()>`
 
 static B: A<()> = todo!();
-//~^ ERROR cycle occurred during layout computation
 
 fn main() {}
