+++ tests/ui/traits/issue-105231.rs
@@ -6,5 +6,4 @@ struct B<T>(A<A<T>>);
 trait Foo {}
 impl<T> Foo for T where T: Send {}
 impl Foo for B<u8> {}
-//~^ ERROR conflicting implementations of trait `Foo` for type `B<u8>`
 fn main() {}
